feat: quickstart in swift is now runnable
