+++ docs/cookbook/quickstart.md
@@ -13,7 +13,10 @@ This page provides an overview of how BDK can be leveraged to create and sync a 
 
 === "Swift"
 
-    Create a new Swift project in Xcode.
+    ```shell
+    swift package init --type executable
+    ```
+    Or, if you're building an iOS app, create a new Swift project in Xcode.
 
 === "Kotlin"
 
@@ -29,8 +32,14 @@ This page provides an overview of how BDK can be leveraged to create and sync a 
 
 === "Swift"
 
+    ```toml title="Package.swift"
+    --8<-- "examples/swift/quickstart/Package.swift"
+    ```
+    Or, if you're building an iOS app:
+    
     1. From the Xcode File menu, select Add Package Dependencies...
-    2. Enter `https://github.com/bitcoindevkit/bdk-swift` into the package repository URL text field
+    2. Enter `https://github.com/bitcoindevkit/bdk-swift` into the package repository URL search field and bdk-swift should come up
+    3. For the Dependency Rule select `Exact Version`, enter the version number (same as Package.swift) and click Add Package
 
 === "Kotlin"
 
@@ -49,8 +58,8 @@ We'll give a breakdown of the key pieces of this code in the next section.
 
 === "Swift"
 
-    ```swift title="examples/swift/quickstart/main.swift"
-    --8<-- "examples/swift/quickstart/main.swift:file"
+    ```swift title="examples/swift/quickstart/Sources/main.swift"
+    --8<-- "examples/swift/quickstart/Sources/main.swift:file"
     ```
 
 === "Kotlin"
@@ -72,7 +81,10 @@ The wallet will take a few seconds to sync, then you should see the wallet balan
 
 === "Swift"
 
-    Run the project in Xcode.
+    ```shell
+    swift run
+    ```
+    Or run the project in Xcode.
 
 === "Kotlin"
 
@@ -93,7 +105,7 @@ First we need some <a href="https://github.com/bitcoin/bitcoin/blob/master/doc/d
 === "Swift"
 
     ```swift
-    --8<-- "examples/swift/quickstart/main.swift:descriptors"
+    --8<-- "examples/swift/quickstart/Sources/main.swift:descriptors"
     ```
 
 === "Kotlin"
@@ -116,7 +128,7 @@ This example is using an <a href="https://github.com/Blockstream/esplora" target
 === "Swift"
 
     ```swift
-    --8<-- "examples/swift/quickstart/main.swift:client"
+    --8<-- "examples/swift/quickstart/Sources/main.swift:client"
     ```
 
 === "Kotlin"
@@ -142,7 +154,7 @@ Once we have our wallet setup and connected to the network, we scan the network 
 === "Swift"
 
     ```swift
-    --8<-- "examples/swift/quickstart/main.swift:scan"
+    --8<-- "examples/swift/quickstart/Sources/main.swift:scan"
     ```
 
 === "Kotlin"
